fix(multiplayer) Added tests for snatching control on a click.
