+++ editor/src/components/editor/store/collaborative-editing.ts
@@ -1,6 +1,5 @@
 import {
   deleteFileFromCollaboration,
-  showToast,
   updateCodeFromCollaborationUpdate,
   updateExportsDetailFromCollaborationUpdate,
   updateImportsFromCollaborationUpdate,
@@ -36,7 +35,6 @@ import {
   TopLevelElementKeepDeepEquality,
 } from './store-deep-equality-instances'
 import type { WriteProjectFileChange } from './vscode-changes'
-import { v4 as UUID } from 'uuid'
 import {
   deletePathChange,
   ensureDirectoryExistsChange,
@@ -48,11 +46,8 @@ import { isFeatureEnabled } from '../../../utils/feature-switches'
 import type { KeepDeepEqualityCall } from '../../../utils/deep-equality'
 import type { MapLike } from 'typescript'
 import { Y } from '../../../core/shared/yjs'
-import { HEADERS, MODE } from '../../../common/server'
-import { UTOPIA_BACKEND } from '../../../common/env-vars'
 import type { ProjectServerState } from './project-server-state'
 import { Substores, useEditorState } from './store-hook'
-import { notice } from '../../common/notice'
 
 const CodeKey = 'code'
 const TopLevelElementsKey = 'topLevelElements'
@@ -613,163 +608,6 @@ function synchroniseParseSuccessToCollabFile(
     ImportsKey,
     ImportDetailsKeepDeepEquality,
   )
-}
-
-export interface ClaimProjectControl {
-  type: 'CLAIM_PROJECT_CONTROL'
-  projectID: string
-  collaborationEditor: string
-}
-
-export function claimProjectControl(
-  projectID: string,
-  collaborationEditor: string,
-): ClaimProjectControl {
-  return {
-    type: 'CLAIM_PROJECT_CONTROL',
-    projectID: projectID,
-    collaborationEditor: collaborationEditor,
-  }
-}
-
-export interface SnatchProjectControl {
-  type: 'SNATCH_PROJECT_CONTROL'
-  projectID: string
-  collaborationEditor: string
-}
-
-export function snatchProjectControl(
-  projectID: string,
-  collaborationEditor: string,
-): SnatchProjectControl {
-  return {
-    type: 'SNATCH_PROJECT_CONTROL',
-    projectID: projectID,
-    collaborationEditor: collaborationEditor,
-  }
-}
-
-export interface ReleaseProjectControl {
-  type: 'RELEASE_PROJECT_CONTROL'
-  projectID: string
-  collaborationEditor: string
-}
-
-export function releaseProjectControl(
-  projectID: string,
-  collaborationEditor: string,
-): ReleaseProjectControl {
-  return {
-    type: 'RELEASE_PROJECT_CONTROL',
-    projectID: projectID,
-    collaborationEditor: collaborationEditor,
-  }
-}
-
-export interface ClearAllOfCollaboratorsControl {
-  type: 'CLEAR_ALL_OF_COLLABORATORS_CONTROL'
-  collaborationEditor: string
-}
-
-export function clearAllOfCollaboratorsControl(
-  collaborationEditor: string,
-): ClearAllOfCollaboratorsControl {
-  return {
-    type: 'CLEAR_ALL_OF_COLLABORATORS_CONTROL',
-    collaborationEditor: collaborationEditor,
-  }
-}
-
-export type CollaborationRequest =
-  | ClaimProjectControl
-  | SnatchProjectControl
-  | ReleaseProjectControl
-  | ClearAllOfCollaboratorsControl
-
-export interface RequestProjectControlResult {
-  type: 'REQUEST_PROJECT_CONTROL_RESULT'
-  successfullyClaimed: boolean
-}
-
-export interface ReleaseControlResponse {
-  type: 'RELEASE_CONTROL_RESULT'
-}
-
-export type CollaborationResponse = RequestProjectControlResult | ReleaseControlResponse
-
-const collaborationEditor = UUID()
-
-async function callCollaborationEndpoint(
-  request: CollaborationRequest,
-): Promise<CollaborationResponse> {
-  const url = `${UTOPIA_BACKEND}collaboration`
-  const response = await fetch(url, {
-    method: 'PUT',
-    credentials: 'include',
-    headers: HEADERS,
-    mode: MODE,
-    body: JSON.stringify(request),
-  })
-  if (response.ok) {
-    return response.json()
-  } else {
-    throw new Error(`server responded with ${response.status} ${response.statusText}`)
-  }
-}
-
-export async function claimControlOverProject(projectID: string | null): Promise<boolean | null> {
-  if (projectID == null || !isFeatureEnabled('Baton Passing For Control')) {
-    return null
-  }
-
-  const request = claimProjectControl(projectID, collaborationEditor)
-  const response = await callCollaborationEndpoint(request)
-  if (response.type === 'REQUEST_PROJECT_CONTROL_RESULT') {
-    return response.successfullyClaimed
-  } else {
-    throw new Error(`Unexpected response: ${JSON.stringify(response)}`)
-  }
-}
-
-export async function snatchControlOverProject(projectID: string | null): Promise<boolean | null> {
-  if (projectID == null || !isFeatureEnabled('Baton Passing For Control')) {
-    return null
-  }
-
-  const request = snatchProjectControl(projectID, collaborationEditor)
-  const response = await callCollaborationEndpoint(request)
-  if (response.type === 'REQUEST_PROJECT_CONTROL_RESULT') {
-    return response.successfullyClaimed
-  } else {
-    throw new Error(`Unexpected response: ${JSON.stringify(response)}`)
-  }
-}
-
-export async function releaseControlOverProject(projectID: string | null): Promise<void> {
-  if (projectID == null || !isFeatureEnabled('Baton Passing For Control')) {
-    return
-  }
-
-  const request = releaseProjectControl(projectID, collaborationEditor)
-  const response = await callCollaborationEndpoint(request)
-  if (response.type !== 'RELEASE_CONTROL_RESULT') {
-    throw new Error(`Unexpected response: ${JSON.stringify(response)}`)
-  }
-}
-
-export function displayControlErrorToast(dispatch: EditorDispatch, message: string): void {
-  dispatch([showToast(notice(message, 'ERROR', false, 'control-error'))])
-}
-
-export async function clearAllControlFromThisEditor(): Promise<void> {
-  if (isFeatureEnabled('Baton Passing For Control')) {
-    const request = clearAllOfCollaboratorsControl(collaborationEditor)
-
-    const response = await callCollaborationEndpoint(request)
-    if (response.type !== 'RELEASE_CONTROL_RESULT') {
-      throw new Error(`Unexpected response: ${JSON.stringify(response)}`)
-    }
-  }
 }
 
 export function allowedToEditProject(serverState: ProjectServerState): boolean {
